// project/client/src/app/entity/activity.ts
export class Activity {
    aid: string;
    name: string;
    begintime: Date;
    length: number;
    needp: number;
    numofp: number;
    siteid: string;
    sitename: string;
    sponsor: string;
    hostid: number;
    introduce: string;
    money: number;
    state: number;
}